add record's communities new resource stubs
* adds a new resource /api/records/<pid>/communities
* creates stub methods
* extracts IIIF to a dedicated service
* extracts /api/communities/<pid>/records to a dedicated resource
