+++ invenio_rdm_records/services/communities/__init__.py
@@ -1,0 +1,12 @@
+# -*- coding: utf-8 -*-
+#
+# Copyright (C) 2023 CERN.
+#
+# Invenio-RDM-Records is free software; you can redistribute it and/or modify
+# it under the terms of the MIT License; see LICENSE file for more details.
+
+"""Record communities service."""
+
+from .service import RecordCommunitiesService
+
+__all__ = ("RecordCommunitiesService",)
