+++ invenio_rdm_records/services/iiif/__init__.py
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+#
+# Copyright (C) 2023 CERN.
+# Copyright (C) 2022 Universität Hamburg.
+#
+# Invenio-RDM-Records is free software; you can redistribute it and/or modify
+# it under the terms of the MIT License; see LICENSE file for more details.
+
+"""IIIF service."""
+
+from .service import IIIFService
+
+__all__ = ("IIIFService",)
