+++ setup.cfg
@@ -74,6 +74,7 @@ invenio_base.api_blueprints =
     invenio_rdm_records_ext = invenio_rdm_records.views:blueprint
     invenio_rdm_records_parent_links = invenio_rdm_records.views:create_parent_record_links_bp
     invenio_rdm_records_record_files = invenio_rdm_records.views:create_record_files_bp
+    invenio_rdm_community_records = invenio_rdm_records.views:create_community_records_bp
     invenio_oaipmh_server = invenio_rdm_records.views:create_oaipmh_server_blueprint_from_app
     invenio_iiif = invenio_rdm_records.views:create_iiif_bp
 invenio_base.blueprints =
